Create README.md
Hi there! I'm Gaurav Chaudhary, a student with a passion for coding. I'm excited to be a part of the GitHub community and to collaborate with other developers on interesting and challenging projects. I'm always eager to learn and improve my skills. When I'm not coding, you can find me in the play ground or on netflix. Feel free to check out my repositories and connect with me on GitHub!